requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9

# Notion API
notion-client==2.2.1

# Authentication & Security
django-two-factor-auth==1.15.3
python-decouple==3.8
django-ratelimit==4.1.0
django-cors-headers==4.3.1

# File Storage
boto3==1.34.0
django-storages==1.14.2
Pillow==10.1.0

# Task Queue (if needed)
celery==5.3.4
redis==5.0.1

# API & Serialization
django-filter==23.5

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0

# Deployment & Production
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Monitoring & Logging
sentry-sdk==1.39.1
django-debug-toolbar==4.2.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
requests==2.31.0

# Security
pyotp==2.9.0  # For MFA
qrcode==7.4.2  # For MFA QR codes

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==2023.10

# Email
django-anymail==10.2  # For transactional emails

# Search (if needed)
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0